requirements.txt
paho-mqtt
langchain
openai
python-dotenv
requests
langchain-ollama 
pandas
streamlit 
matplotlib
langchain-community
